requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-CORS==3.0.10
Flask-RESTx==0.5.1
Flask-Limiter==2.4.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
SQLAlchemy==2.0.28
alembic==1.13.1
redis==4.0.2
python-dotenv==1.0.1
PyJWT==2.1.0
prometheus-client==0.11.0
requests==2.31.0
python-dateutil==2.8.2
Werkzeug==3.0.1
geopy==2.2.0
pdfkit==1.0.0 
Flask-JWT-Extended==4.6.0
apscheduler==3.10.4
openai==1.3.0
pytz==2024.1
fpdf==1.7.2
email-validator==2.1.0.post1
pandas==2.0.3
xlsxwriter==3.1.0
reportlab==4.1.0
cryptography==42.0.2 